Update sound.py (#230)
* Update sound.py

Issue: #229

Add a Player class to make management of background music easier.

* Update sound.py

Issue: #229
Fixed test parameters, forgot to import arcade.

* Update sound.py

Changed how the Player handles audio to use the arcade functions for audio.
Changed the audio to be loaded statically so that it is a one and done thing.
Changed how the Player handles audio so that each track gets its own player to speed up playing.
Changed how playing works so now all that is required is to load file/directory then play the track by name.
Added ability to resume on play or reset when played allowing for some interesting interactions.
Added ability to play more than one track at a time through a parameter being true/false.
Added ability to stop single track by name.

Overall the usage of the Player has been simplified allowing for more complex interactions. As a result the Player can support multiple background tracks simultaneously allowing for complex patterns to emerge if desired.

* Update sound.py

Bug in player.volume was checking if it was not a pyglet.media.Player object leading to it trying to change the volume of a string.