Fix Vec2d.__iadd__ bug
The position setter has a check to make sure the new position is actually different from the current position (to prevent re-hashing the sprite location).

The problem was, when using `Vec2d.__iadd__`, E.g.,  `sprite.position += (2, 2)`, the Vec2d was changing the vector in place, and then sending that changed (inplace) vector to the position.setter where it would check to see if the position was changed. The new_vector sent to the setter and the `self._position` were always the same, since the `Vec2d.__iadd__` already changed `self._position` in-place.

The fix was to also check if the sprite has a relevant velocity; this is, a velocity whose magnitude is greater than 0.001.