Add sprite._position_cached
Fix for Vec2d in-place methods which were changing the position vector before the position setter can check if it needs to recalculate the spatial hash. Because it changed the positon before that, the `Sprite.position` setter was never detecting a difference between the new vector and the current one.