Window commands cleanup (#1538)
* Remove clear_window() (unambiguous)

* Remove redundant get_pixel_ratio

* Remove orthogonal_projection()

This function predates pyglet.math.Mat4.orthogonal_projection().
We should use that instead. Also this is a low level function
that should not be in the arcade namespace.

* Fix missing near/far in atlas resize projection

* Group together start/finish_render

* Remove arcade.get_projection

This is not a compatible pair with get_projection and can
cause a lot of confusion. The window already has a projection
member / function. Use that instead.

* Remove get_viewport

* Remove get_projection

* Remove unused import