Improve robustness of the line of sight function
Previously, for the case where the `distance` between the `observer` and `target` is zero the function will crash with a Division by Zero Error.
This error can happen if someone designs a chasing enemy that ends up at the players location.
I added a check will return True indicating there is line of sight in such a case.

Perhaps this should instead be a configurable option?

Additionally, the Line of Sight function did not ensure that we have a positive non-zero value for the check resolution.
I added a check that will raise a value error in such a case.

Also added unit test for the same position for LOS