Added the properties for platforms and walls a la ladders (#1972)
* added the properties

converted platform init check into a one line so that it is in line with the other checks

* fixed deleter properties for platforms/walls

* added property to the simple engine so wall list can be updated on it

* Removed old imports 

BasicSprite type was removed from simple physics engine. Only statement that used it was the engine's wall list init setter statement. Where it limited the spritelist's contents to basic sprites only. Typing.cast was also removed as it seemingly doesn't do anything that explicitly setting the type.