Create a BillboardList subclass of SpriteList
BillboardLists make sprites always face the screen. This is useful for something like isometric. Where the ground tiles are flat, but walls and other decorations should still face the camera.

Sprites assume they are on the x-y plane a lot so all of the left, right, bottom, top methods aren't accurate for a sprite that is drawn with a billboardlist.