Updated scale method names, and reimplimented scalar setting of scale
The names we had for the scale methods didn't feel right for arcade/python so they have been changed to be more readable.

While the concern about confusing the user is valid I think the change is worth it. During the interim before we do a larger re-write this is a nicer developer experience than using a method. Most developers set the sprite scale once, and then don't change it in the future.

the `isinstance` check for `AsFloat` is no-where near the cost of creating a vector.

In my ideal world sprite would not have a scale property at all, and instead scale is a method that changes the size of the sprite.